tls/ossl: Adjusted value of ERROR_UNEXPECTED_PACKET
Now using SSL_R_WRONG_VERSION_NUMBER, which is the error OpenSSL returns
when it receives a non-TLS packet when expecting a TLS handshake packet.
This helps Redpanda because it treats ERROR_UNEXPECTED_PACKET as a
benign disconnect error.

Added method that can be used to override the OpenSSL message with one
that may be more helpful to users.

Signed-off-by: Michael Boquard <michael@redpanda.com>